Adjust minimum word count for judge feedback form
This will change the minimum word count for the "Overall Impression"
section to 40, and for all other sections it will be 20.

I adjusted colorForWordCount() to accommodate for the different/dynamic
minimum word counts as well. I also updated one of the colors,
darkyellow, b/c it isn't a valid named color.

https://developer.mozilla.org/en-US/docs/Web/CSS/color_value

Refs: #2444