Add link to student dashboard to view submission
I added a new guard to control which submission link to display,
thanks for the suggestion @stenington!

It's not the typical early-return guard. It's more like a "Hey guard, can I
display the XXX submission link?" type of guard; a guard to control which
link can be displayed.

This new guard is used on the student's dashboard to determine whether
or not to display a link to view the team's completed submission.

Refs: #2517