Add faux account
Since parents don't have accounts on the platform, we will create a faux
account for them when they subscribe to an email list or when the need
updating on an email list.

It's called `FauxAccount`, but really only the naming is preventing it
from being just a generic faux object.

Refs: #2710