Fix error when mentors search for other mentors
The `_mentor_list.html.erb` partial is shared by both students and
mentors (and admin and chapter ambassadors!).

The bug was with the link to the mentor search page - students have a
different version than mentors.

I updated the link to fix the bug, I also added a basic spec to check
for the correct link.

Refs: #2836