Handle multiple location suggestions during registration
It seems that during registration, when multiple locations are returned
from geocoder, you can click on one and submit, but then it just does
another geocoder search which returns multiple results... and then it
just keeps repeating that loop.

This change will make is so that when you click a suggestion, that will
get saved as your location.

Refs: #2845