Update old migrations to handle new soft-delete functionality
Old data migrations that reference student profiles were failing b/c
the newly introduced soft-delete functionality exists in the model, but
the `deleted_at` column doesn't exist in the database (at the time when
these migrations run).

The solution was just to redefine the student profile model in the
migrations to not include the soft-delete functionality.

This is the approach I took:
https://railsguides.net/change-data-in-migrations-like-a-boss/#redefine-models-in-migrations

Refs: #2354