Fix mentor who can't be hard-deleted
The consent waiver needs to be deleted before the account or else it
will throw a "PG::ForeignKeyViolation" error, b/c it will try to delete
the account but an associated consent waiver record will still exist.

I was able to confirm this fix by pulling down the production database and
testing this fix on the account that was having issues. But very
weirdly, I can't duplicate the issue locally w/ QA data. 😕

Refs: #2848