Require images for team submission
Notes:

The tests don't actually upload any images (formerly known as a
screenshots), a lot of specs were failing b/c the views want to display
an actual image via a url, well these images don't exist for the tests,
doh! I tried to use/attach real images for our tests, but just couldn't
get it working. 😥 So I just added a conditional to only display images
if the url for them is present.

The error I was running into, and the solution (which I thought: that's
pretty simple!):
https://stackoverflow.com/questions/53510040/carrierwave-argument-error-nil-location-provided-cant-build-uri-for-an-image

Refs: #2850