Reload page after uploading an image
This will force a page reload after an image has been uploaded. This is
needed for when images are the last piece for a complete submission; the
page needs to be refreshed in order for the submit button to be enabled,
without this fix you would have to navigate to another section/page in
order for the submit button to be enabled.

Refs: #2850