Update judge's finished scores section
The router and "nav" components were expanding/collapsing each section
based on which route was active.

We don't want this expand/collapse functionality, we just want to
display all the score, so the router and "nav" components aren't needed
anymore.💥

To have these sections always displayed, I just replaced the "nav" with
the components the "nav" was associated with.

Refs: #2855