Add mentor to judge email list when becoming a judge
When a mentor becomes a judge (via the "Switch to Judge Mode" button),
this will add them to the "judge" email list (by tagging them as a
judge). The account will only get tagged as a judge when the judge
profile is first created (when the button is clicked for the first
time), and only if they are currently on the Mailchimp list.

Linting rule on "IneffectiveAccessModifier":
https://docs.rubocop.org/rubocop/cops_lint.html#lintineffectiveaccessmodifier

Refs: #2912