Update spec debugging settings
This change will only allow "true" in the .env file to turn on the
setting to automatically open a screenshot or page when a spec fails.
"False" (and all other values) in the .env file will disable these
settings.

Before, "false" would actually be true b/c the string "false" is truthy.

Hopefully this change makes it a little clearer.

Refs: #2872