Remove qt5 dependencies
Our CircleCI image was changed from Buster to Bullseye. The qt5-default
package doesn't exist in Bullseye and is therefore breaking the build.

I don't think we need this package, from what I can tell, it's used for
capybara-webkit, which we're no longer using and is no longer being
maintained; so I'm removing this dependency!

https://github.com/thoughtbot/capybara-webkit

https://github.com/Iridescent-CM/technovation-app/commit/d7cd3a815dbeeba10bf090282e5886a236ce4da3#diff-21b1fe66043572c76c549a4fc5f186e9a69c330b186fc91116b9b70a4d047902

Refs: #3028