Add welcome message and required steps for students
I added a "notifications" area in our main layout for notification type
content. The first thing to use this will be a new welcome message
for students that also has required steps in order to get started;
this is a Rails partial, which I'm thinking almost as a Vue SFC, where
everything is contained in there, although it's not really re-usable at
this point.

Also, the links under "Profile Settings" (Basic Information, Parental
Consent, Location, etc.) don't go to pages, they just dynamically
show/hide content. Which means there is no way to link someone to go to
one of those sections, which I wanted to do for the parental consent
for this new notification - like go here for your parental consent. To
do this I had to add some functionality that will check the url for an
anchor and then programmatically click that link which will show the
right content; this will only apply to the parental consent link, I
didn't feel comfortable adding it for all links, I don't know if that
would introduce other weirdness since Vue uses hash mode routing.

Refs: #3072