Fix dynamic team url
I think there's some Rails magic happening here - I'm pretty sure
`team_preview` is b/c the partial is named that - and a team gets passed
in to the partial - and the team becomes the team preview.

Anywho, we want the id of the team/team preview for the url param, which
is what I added.

Refs: #3190