Update code completion logic
The submission flow uses a bit of magic where everything has be named
just right for it to work correctly.

For a section to be marked complete, it checks for a certain
TeamSubmission attribute, and if it's not complete then it will load
the corresponding page - this page has to be named exactly like the
attribute.

We added a `submission_type` field, but left the page named
`development_platform', this was causing things to malfunction, to fix
this I added a check explicitly for `submission_type` and then set the
attribute back to `development_platform` so things will work like they
were before.

When we get around to properly renaming everything, this workaround can
be removed.

Refs: #3011