Add new submission review page to submission flow
Right now, the only way to view the submission review page is on the
dashboard after a submission has been completed.

This change will make it so that the submission review page will also be
viewable as a part of the submission flow - it will be shown as the last
step right before the submission is submitted/finalized.

Regarding the tests, the new submission review page doesn't display a
Thunkable URL, which the tests are expecting, so that's why those tests
were removed.

Refs: #3229