Filter demo questions based on submission type
There are six possible questions for the demo section - three for AI
projects and three for Mobile App projects.

Prior to this update the back-end would have returned all six questions
to the front-end.

With this update the back-end will only return the necessary questions
based on whether the submission type is an AI Project or a Mobile App
project. This will make things simpler for the front-end.

Refs: #3342