Wire up score stepper
This was a very tricky issue to resolve, so it's tricky to describe.
Basically, fwict, the score stepper won't work unless it's outside of
the "judge-scores-app", which is the way it worked before, I restored
this functionality the best I could. I had to add some extra closing
`div` tags in order for the layout to display correctly. I'm not sure
why that is necessary yet, I think it suggests that are `<div>` tags
are misaligned somewhere.

Refs: #3342