Redirect to homepage if registration is closed
This will make is so that if someone clicks or visits the signups URL,
then it will redirect them to the homepage. Which means they won't ever
see the new registration flow, which mean the new registration doesn't
need to know if registration is closed (so I removed the code related to
that).

Refs: #3255