Add admin controls for judge registrations
This will give admins the ability to toggle judge registrations on and
off.

Unlike student and mentor registrations, judge registrations will be
independent of whether judging is turned on or off i.e. student and
mentor registrations can't be enabled if judging is turned on, but judge
registrations can be turned on or off regardless if judging is turned on
or not.

Also, note that this is only the admin controls, coming soon will be to
incorporate this new setting into the new registration.

Refs: #3414