Update total score functionality
The demo section has different questions depending on the submission
type - there are three questions for a mobile app submission, and there
are three questions for an AI submission. The total score functionality
had a bug that calculated the total score based on six questions, when
really there are only three. I updated the code to factor in the `idx`
value which contains unique values for each section.

Refs: #3546