Update CircleCI to prepare for Ruby 3.1.2 upgrade
In order to upgrade to Ruby 3.1.2, CircleCI had to be updated to use the
new next-generation convenience images - we are currently using the
legacy images, which doesn't have an image supporting Ruby 3.1.2.

Here is some material on legacy and next-gen convenience images:
https://circleci.com/docs/circleci-images

Our old image was based on Debian, the new one is based on Ubuntu, this
caused two of our packages to break - PostgreSQL and wkhtmltopdf.

I also had to update the way browser tools are installed, I followed the
steps here:
https://circleci.com/developer/orbs/orb/circleci/browser-tools

Lastly, regarding the Ruby 3.1.2 upgrade, this commit will not be
complete/working until all of the commits in this branch get merged in.
I just wanted to break down the commits to more reasonable chunks.

Refs: #3575