Update survey link spec
This is one of the tests that was randomly failing b/c of the CircleCI
vars. I'm pretty sure it was failing b/c in the tests we're doing 2
weeks + 2 weeks, and there is supposed to be a survey reminder link, but
if we're close (4 weeks) to the next season then there won't be a survey
link b/c the platform will think it it's the next season as defined by
the CircleCI vars.

I updated the tests to be 2 days instead of 2 weeks b/c back in 2018 the
production code was updated from 2 weeks to 2 days, but the tests were
never updated.

bcd37513cf

In the future, I'm pretty sure this will still be an issue, except that
it won't an issue for a four week window, it will just be an issue for a
four day window.