Update "health" question for submissions
The "health" question was updated to be a hunger/food waste question.

This is only a bare minimum change to get the question updated,
the code/back-end is pretty confusing though b/c this question was
originally for a "game" question, then a "health" question, and now a
"hunger" question, and in the code there are remnants of these old
versions which makes it confusing! Hopefully we can go back and clean
this up!

Refs: #3713