Add migration to fix beginner -> junior students
Beginner students are allowed to have their account email address be the
same as the parent/guardian email address.

Junior students can't have their account email address be the same as
the parent/guardian email address.

Returning beginner students who will be juniors in the new season
violate this email validation behavior^.

This migration will find all the affected students for the 2022 season
and update the parent/guardian email address to be `nil`.

Refs: #3771