Default season to the current season for judge questions
A pretty simple change that will default the season to the current
season if one wasn't provided when trying to fetch the questions.

There is RPE functionality that allows someone to view the scores for an
RPE event, but if a judge hasn't scored a submission yet, then the score
will be new/blank without a season, which leads to an error b/c it can't
find the questions if a season isn't specified.

Refs: #3792