Update profile updating order
Since we're now surfacing two errors when the account email address is
the same as the parent/guardian's - one on the `account` and one on
`student_profile` (this one is new), this will change where the
`profile_controller` is redirecting to.

I moved the `elsif profile.errors["parent_guardian_email"].any?` section
below the `elsif profile.errors["account.email"].any?` to preserve the
existing behavior (so 'email_addresses/edit' will get rendered).

Refs: #3809