Add judging section name lookup functionality
Updating the existing `sections_for` method to return a hash
(with section and section display name) instead of just an array
(with on of the values).

Part of this new functionality is the `section_display_name_for` method
which will convert the judging section to a display name, e.g. ideation
-> "Learning Journey".

This will also update existing score section functionality since
`sections_for` now returns a hash instead of an array.

Refs: #3804, #3386