Update background report checking
The main change here is to accommodate for an "engaged" adjudication
status. If a candidate goes into the "consider" state, someone can log
in to Checkr and click "Engage" for this candidate, if this happens we
want to mark this background check for this candidate as "clear". This
is existing behavior that needed to be moved/adjusted for the recent
Checkr API changes.

- https://info.checkr.com/rs/204-WFI-586/images/Report%20Lifecycle%20Design%20Decisions%20Checklist.pdf

I also added some logging statements to keep the new additions
consistent with what we are doing now.

Refs: #3793