Add "not started" scores functionality
This will collect "not started" scores and return them with the scores
response. Then the "not started" scores will be displayed in the
"Submissions to Score" table.

"Not started" scores is a combination of assigned submission via RPEs
and scores that haven't been started i.e. scores where `created_at` is
the same as `updated_at`.

The assigned submissions functionality was moved from the
`AssignedSubmissionsController` and slightly tweaked.

Refs: #3869