Update sign_out helper
I think it makes more sense for this helper to visit the signout_path vs
looking for a sign out link. I think it makes it too brittle by looking for a
link. It should be a helper to sign out and not to verify the sign out
link itself, imho.

Refs: #3854