Update survey specs
ImportantDates.new_season_switch has to be hard coded to a date in the
past.

If this date is not hard coded, it will use the ENV vars
(DATES_NEW_SEASON_START_MONTH and DATES_NEW_SEASON_START_DAY), and when
the specs run within 2 days of what's defined in the ENVs these specs
will fail! Because the survey will get created in one season, the season
will advance to the next season and there won't be a survey reminder for
the old survey (which is what these specs are testing for).

Refs: #3860