Add new video clicked tracking functionality
For some reason our existing video tracking isn't working, I have a
hunch it's b/c the pop-up modal is intercepting/blocking the click that
makes the request to the back-end to update the video tracking fields
in the database. Anywho, I added some new functionality that doesn't
require an extra HTTP request but instead hooks into the "embeds"
controller, and looks at what is being embedded (either the pitch or
demo/technical video) and updates the tracking field (for the specific
submission score) in the database accordingly.

Refs: #3971