Add tests for LearnWorlds functionality
Just some really basic/general test cases here.

I know it's not complete, like for example there are no tests for
testing the auto-enrollment functionality. But writing the tests have
been a slog and there's a chance we might not use it next season so,
probably not worth spending more time on it at the moment.

Refs: #3192