Chore: Address judge registration invites w/o a name
When a chapter ambassador invites a judge they just provide the judge's
email address and not their name; name used to be defaulted to "Invited
Judge", but that default has been removed so that name can be used when
an admin invites someone to register. The changes here are to account
for when judges don't have a name.

Refs: #4216