Chore: Ensure ChA registration invites are "any time" invites
When an admin invites a chapter ambassador, the GUI will pre-check the
"register at any time" checkbox and disable it. The issue is that a
disabled field won't get submitted to the server, so it's not actually
creating an any time invite.

This change will address this by updating the corresponding
attribute in the controller, and hard-coding the response from the
`RegistrationInviteCodeValidator` service.

Refs: #4296