Feature: Limit number of recusals a judge can preform
This feature will limit the number of recucals a judge can preform.
The limit is defined in a new configuration variable named,
`JUDGE_MAXIMUM_NUMBER_OF_RECUSALS`. If a judge exceeds this limit we
will display a message to them indicating this and to contact us.

This feature should help with the issue where judges continuously recuse
themselves until they find a particular submission.

Refs: #4320 #AmeriCorps