Chore: Update specs for new "Chapter Admin Activity"
When a ChA signs in they get redirected to their dashboard, the problem
is that we have created a new ChA dashboard for onboarding. And these
specs are expecting to see the old dashboard, which is now referred
to as "Chapter Admin Activity". So, the specs need to be updated
accordingly to visit the `chapter_ambassador_chapter_admin_path` (aka
the "Chapter Admin Activity" section, which was their old dashboard).

Refs: #4392