UI: Add active style for sub nav items
This will add a style for when a sub nav item is the active page. The
code and styles were already in place, but it wasn't working correctly,
I fixed it to work properly. We can update the styling later, but for
now at least it is a starting point.

Refs: #4411