Fix: Axios CSRF issue
Recently axios was updated to address a security vulnerability
and some things changed/broke with that. This update will change
`config.headers.common` to `config.headers`, which will properly include
the CSRF token in the headers with axios calls. The rest of the changes
in this file are formatting updates.

Refs: #4477 #4485 #4498