UI: Add sidenav to ChA background check flow
This will make the sidenav visible for the entire background check flow.

I created a new partial
`chapter_ambassador/dashboards/onboarding/sidenav` that can be included
whereever we want show the sidebar.

I updated the route for chapter ambassador background check show page to
be a singular resource which will remove the id from the URL.

Refs: #4495