Chore: Add rake task to migrate ChAs to chapters
This will add a rake task that can be used to convert chapter
ambassadors to chapters.

It works by passing the rake task a list of pairs of data - chapter
ambassador ids and the organization name, separated by a semicolon; and
each pair of data is separated by a comma.

```
bundle exec rake convert_chapter_ambassadors_to_chapters["ID;ORG NAME,ID;ANOTHER ORG"]
```

If there are commas and semicolons in the organization name they would
need to be escaped.

Refs: #4442