Fix: Parental consent status bug
I'm not sure how, but it's possible that a student can have multiple
parental consent forms for a season. Our app cannot handle this if this
happens, the participants datagrid can display a different result than
the student profile debugging section, one could display signed and the
other pending.

To fix this, I created one source, `parental_consent_signed?` to check
if *any* of the parental consents for the season are signed, this method
will be used in both the datagrid and the student profile debugging
section.

Refs: #4492