Chore: Update chapter details page
A very basic update here to add chapter name, location and summary to
the chapter's show page. The layout will basically be the same as the
participant's show page.

On the participant's show page the location is shown via
`#address_details`, in order to get the same functionality for chapters,
I had to add `include ActiveGeocoded` to the chapter's model.

Refs: #4562