Chore: Update meets_minimum_age_requirement migrations
This should ensure that this migration will only run if the
`meets_minimum_age_requirement` column doesn't exist on QA or Preview.
I'm pretty sure this is needed because both QA and Preview have
this migration (one was added on QA for judges and one for chapter
ambassadors on Preview), and this should ensure the column will only be
added one time.

Refs: #4537