Chore: Add host to Saleforce API calls
By default, if a host is not specified, Restforce will use the
production host (`login.salesforce.com`), for the sandbox we need to use
`test.salesforce.com`, which needs to be specified. The changes here
will allow us to specify which host to use so we can set it for the
sandbox environment.

https://github.com/restforce/restforce?tab=readme-ov-file#sandbox-organizations

Refs: #4386