Chore: Refactor chapter legal contacts
This will move legal contacts to a separate table, and make name and
email address required fields, which will ensure that name and email
address will be present when we send a legal contact the legal document
to sign via DocuSign. This will also make all the necessary changes so
that an admin can add and edit a legal contact like before.

Refs: #4691 #4649