Chore: Update DocuSign service to accept a legal contact
This will update the DocuSign service, job, and specs to make it so that
a legal contact can be passed to the service, instead of individual
fields (full name, email address, job title, organization name). By
using a legal contact, it will ensure that the full name and email
address are present since those are required fields and a legal contact
cannot exist without them.

Refs: #4691