Security: Update nanoid node module
This will update all versions of nanoid to version `3.3.7` to address
this security vulnerability:

- https://github.com/Iridescent-CM/technovation-app/security/dependabot/70

To address this vulnerability, nanoid has to be updated to at least
version `3.1.31`.

When I updated Vue Formulate last week that only removed nanoid version
`2.1.11`, but we were still using version `3.1.28` which is still
vulnerable.

The changes here will make sure we're only using version `3.3.7` which
should address the vulnerability.

Refs: #4720