Chore: Update judge recusal functionality
This will make it so that if a score is marked complete and then a judge
tries to recuse themselves from that score, it will mark the score as
incomplete and unapproved. I think this might be possible if a judge has
the score open in two windows/tabs and completes it in one and recuses
in another, another possible way (which this fix one address) is if the
judge recuses themselves from a score and they get reassigned the same
submission later.

Refs: #4732