Chore: Refactor DocuSign API client to be more flexible
I extracted the common parts for making an API call to DocuSign into a
new private method called `send_document_to`, which takes a `signer`
and `params` (which contains all of the specific details). All of the
specific details for sending an MOU to a legal contact have also been
extracted, and this will get passed to `send_document_to`. These changes
will allow for chapter ambassador profiles to be more easily used in
this service.

Refs: #4697