Fix: Saved "Scores exports" bug
A saved "scores export" would go to the scores page instead of the score
exports page; this will fix that issue.

Saved searches are saved with the type of datagrid it is using:

- The "scores export" page uses the scores_grid
- The "scores" page uses the scored_submissions_grid

Both types of these saved searches were being associated to the "scores"
page. The change here will make a saved "scores export" used the "scores
export page".

Refs: #4747