UI: Add admin view for chapter ambassador onboarding
This will add a new partial that will display the basics for chapter
ambassador onboarding. The onboarding steps are:

- Complete Background Check
- Take Training Quiz
- Sign Legal Agreement
- Review Community Page

Only the background check and legal document are hooked up right now.

For the background check, right now in the admin area we track the
status of the background invitation and the background check separately.
But for chapter ambassador onboarding purposes the background check is
either complete or incomplete. If it's incomplete, we'll display some
additional information below the incomplete status; I tried to capture
all the info that we were displaying before.

Refs: #4743