Feature: Restrict access to Chapter Admin Activity
This will restrict access not only to Chapter Admin Activity, but all
chapter ambassador pages until the chapter ambassador and their chapter
are fully onboarded. Any controller/pages that need to be accessible
without having to be onboarded need to explicitly opt out of this
functionality; that way all existing controllers and future controllers
are covered by default.

Refs: #4689