Chore: Update factories to support onboarded ChA and chapter
There are a lot of specs that test things in the Chapter Admin Activity
area, this area is now restricted; chapter ambassadors and their chapter
must be onboarded to be able to access this area now. The updates here
are to make it so that a default chapter ambassador created via
FactoryBot will be associated to a chapter and be onboarded.

Refs: #4689