Performace: Add index on memberships for member_type
We have two queries (in `Account.rb`) that search the `memberships`
table, one query for mentors (where `member_type` is for a mentor) and
one for students (where `member_type` is for a student). Both of these
queries use a table scan, I added an index on `member_type` to help with
this.

Refs: #4783