Feature: Insert program participant records in Salesforce
When someone gets registered to the current season (either by signing up
for a new account, or by logging in for the first time in a new season),
a program participant record will get added in Salesforce for them.

This initial record will have the season and profile type, and for
students it will also include their division, and for mentors it will
also include their mentor types.

Refs: #4855