Feature: Update student's program info in Salesforce
This will add all the necessary pieces to update a student' program info
in Salesforce, the following fields will get updated:

- Team name
- Submission project link
- Submission pitch video link
- Submission submitted or not

Whenever any of these fields are updated on the platform, it will
trigger a job to update these fields in Salesforce. A team (and
submission) can have multiple students, we will create a job for each
student on the team (or belonging to the submission) to update their
info. If mentors are also on a team being updated, there info will not
get updated (because they don't have this info in Salesforce).

The null team/submission values were updated/added so that blank/clean
values will get sent to Salesforce.

There are also some Standard Ruby formatting fixes here too.

Refs: #4884