Chore: Add basic phone number validation for registration flow
I added a very basic phone number validation. It will still allow
gibberish to be entered, but it will limit what gibberish can be
entered. Also, noting that this regex/validation should probably be
extracted and not duplicated, but there's no time for that right now.
I'll add a new issue to address this and improve the validation.

I got the regex from this Stack Overflow post:
https://stackoverflow.com/a/73054458

Refs: #4877