Chore: Add rake task to make ChAs legal contacts
This rake task will make a chapter ambassador a legal contacts for their
chapter.

This rake task needs a list of account ids, for each account id, if
they are a chapter ambassador, they belong to a chapter, and a legal
contact isn't already setup, it these conditions are met it will make
the chapter ambassador the legal contact for the chapter.

The rake task is run with this command:

`bundle exec rake set_chapter_ambassadors_as_legal_contacts[account_ids]`

Refs: #4934