Chore: Only add Salesforce program participant record if upsert succeeded
This is a band-aid/hack fix where if the upsert to Salesforce fails, we
won't try to add a program participant record because we won't have a
contact id for them (which is needed to create the program participant
record). A real fix is figuring out why the upsert can fail with the
error message:

`DUPLICATE_VALUE: duplicate value found: Platform_Participant_ID__c`

Upsert should update or insert by the `Platform_Participant_Id__c`,
there shouldn't be any duplicate ids in Salesforce because it's marked
as distinct. I don't have a fix for this at the moment, so the "fix"
here is to stop trying to create the program participant record if there
isn't a contact id, which should stop these errors from coming in:

`REQUIRED_FIELD_MISSING: Required fields are missing: [Contact__c]`

Refs: #4904