Chore: Prevent duplicate legal agreements from being sent
This will add a check at the controller level to prevent a duplicate
legal agreement from being sent.

I also added a little hack to hide the "Review and Sign" button,
where we pass a `legal_agreement_sent` param to the view and if that is
present, we will hide the button, this should help with people clicking
the button again (because it won't be displayed).

Refs: #4917 #4947