Fix: Salesforce duplicate error messages
For any combo accounts (e.g. mentor/ChAs), we will space out their
Salesforce calls by 30 seconds.

Duplicate error messages, according to this
[document](https://help.salesforce.com/s/articleView?id=000383480&type=1)
can be caused by:

> You are sending a duplicate record while the other is already present/being processed.

And a fix would be:

> it would be a good practice to program the requests in a way to have some milliseconds of time gap between two requests (if the same record is being sent multiple requests)

So that's what I'm trying here with the 30 second spacing.

I added the conditional because on CircleCI it seems the jobs need to
run "inline", and the default for local environments is also "inline".

Also, in order to get the 30 second spacing to work locally you have to
set the backend to "async":

`ACTIVE_JOB_BACKEND=async`

Refs: #4949