Chore: Add season expires to documents
This will add a new `season_expires` field to our documents table. When
a chapter affiliation agreement is signed we will set this field to the
season the document expires in.

I also added a rake task to set this new `season_expires` field for all
signed chapter affiliation agreements.

Refs: #4968