Chore: Add first and last name validation during registration
This will require at least one alphabetical character be entered for the
first name and last name fields during registration.

The validations are duplicated, ideally they would be extracted to avoid
this duplication, but since we're planning to remove Vue.js in the
future, I didn't want to spend more time than necessary on this.

Refs: #4911