Chore: Update first/last name validation
This will make it so that a first name or last name can't begin with one
or more dashes or periods; dashes and periods are fine anywhere else
though. This is just a start for name validation, it will cover the
known issues we've had, we can definitely add more special characters to
the list and improve it.

Refs: #4911