Chore: Refactor chapter assignments
This will change how chapter assignments are done, a chapter assignment
used to be one-to-one - one profile type assigned to one chapter. This
new way will associate accounts (not profiles) to chapters, and will
allow accounts to be able to be associated to multiple chapters.

Since chapters are now associated to accounts, the other big change is that a
chapter's primary contact will now be associated to an account instead
of a chapter ambassador profile.

The rest of the changes here are to support those two^ changes.

Refs: #5024