Feature: Save chapter assignments
The main change here is that this will allow someone to save their
chapter selection, including the "I'm not sure" option. After their
selection is saved, the force chapter selection screen will no longer be
displayed for them.

A lot of the other changes are to support the saving functionality.

Refs: #5030