Chore: Update season filter for ChA's participants datagrid
This will add a new "season" filter for chapter ambassadors to make it so
that the filter will only return participants who were a part of their
chapter for a given season.

The admins will use the existing/original "season" filter.

In the view I updated `count` to `length` because from the datagrid, the
group by clause returns a hash (e.g. `{account_id_x: 2, account_id_y:
2}`) and we just want to know how many records were returned, i.e. the
length of the hash.

Refs: #5078