Chore: Add edit/update functionality for admin chapter assignments
This will make it so that if a student, mentor, or ChA doesn't have
a chapter assignment it will create one for them via the `create`
action, but if they already have an assignment it will update it via the
`update` action.

I made the edit/assign buttons dynamic for ChaA/mentor account combos;
the admin functionality is designed to only be able to edit one
assignment, the button will reflect if the ChA or mentor's assignment is
getting created or updated.

I also updated and added specs and extracted the form into a partial so
it could be reused.

Refs: #5048